Refactor journalbeat build logic
