+++ x-pack/journalbeat/Makefile
@@ -1,0 +1,4 @@
+#
+# Includes
+#
+include ../../dev-tools/make/xpack.mk
