+++ x-pack/journalbeat/journalbeat.docker.yml
@@ -1,0 +1,11 @@
+journalbeat.inputs:
+- paths: []
+  seek: cursor
+
+processors:
+- add_cloud_metadata: ~
+
+output.elasticsearch:
+  hosts: '${ELASTICSEARCH_HOSTS:elasticsearch:9200}'
+  username: '${ELASTICSEARCH_USERNAME:}'
+  password: '${ELASTICSEARCH_PASSWORD:}'
